requirements.txt
pyudev
yara-python
pyclamd
psutil
rich
